Cast to float before computing mean (#408)
* Cast to float before computing mean

Fixes #407.

* Pass PyTorch tensor to normalization callable

+++ torchio/transforms/transform.py
@@ -398,7 +398,7 @@ class Transform(ABC):
 
     @staticmethod
     def mean(tensor: torch.Tensor) -> torch.Tensor:
-        mask = tensor > tensor.mean()
+        mask = tensor > tensor.float().mean()
         return mask
 
     @staticmethod
